+++ lib/hanami/utils/callbacks.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require "concurrent/array"
+
 module Hanami
   module Utils
     # Before and After callbacks
@@ -18,7 +20,7 @@ module Hanami
         #
         # @since 0.2.0
         def initialize
-          @chain = []
+          @chain = Concurrent::Array.new
         end
 
         # Appends the given callbacks to the end of the chain.
@@ -150,6 +152,17 @@ module Hanami
           end
         end
 
+        # Return a duplicate callbacks chain
+        #
+        # @return [Hanami::Utils::Callbacks] the duplicated chain
+        #
+        # @since 2.0.0
+        def dup
+          super.tap do |instance|
+            instance.instance_variable_set(:@chain, instance.chain.dup)
+          end
+        end
+
         # It freezes the object by preventing further modifications.
         #
         # @since 0.2.0
@@ -169,6 +182,10 @@ module Hanami
           super
           @chain.freeze
         end
+
+        protected
+
+        attr_reader :chain
 
         private