+++ spec/unit/hanami/utils/class_attribute_spec.rb
@@ -6,7 +6,7 @@ RSpec.describe Hanami::Utils::ClassAttribute do
   before do
     class ClassAttributeTest
       include Hanami::Utils::ClassAttribute
-      class_attribute :callbacks, :functions, :values
+      class_attribute :callbacks, :functions, :values, :initial
       self.callbacks = [:a]
       self.values    = [1]
     end
@@ -42,6 +42,16 @@ RSpec.describe Hanami::Utils::ClassAttribute do
       self.engines = 2
       self.wheels  = 8
     end
+
+    class DoubleInclude
+      include Hanami::Utils::ClassAttribute
+      class_attribute :foo
+      self.foo = 1
+
+      include Hanami::Utils::ClassAttribute
+      class_attribute :bar
+      self.bar = 2
+    end
   end
 
   after do
@@ -51,7 +61,8 @@ RSpec.describe Hanami::Utils::ClassAttribute do
        Vehicle
        Car
        Airplane
-       SmallAirplane].each do |const|
+       SmallAirplane
+       DoubleInclude].each do |const|
          Object.send :remove_const, const
        end
   end
@@ -68,6 +79,10 @@ RSpec.describe Hanami::Utils::ClassAttribute do
       example.run
 
       $DEBUG = @debug
+    end
+
+    it "the initial value is nil" do
+      expect(ClassAttributeTest.initial).to be(nil)
     end
 
     it "the value it is inherited by subclasses" do
@@ -145,5 +160,13 @@ RSpec.describe Hanami::Utils::ClassAttribute do
     it "doesn't print warnings when it gets inherited" do
       expect { Class.new(Vehicle) }.not_to output.to_stdout
     end
+  end
+
+  it "class_attributes class variable is a Concurrent::Map instance" do
+    expect(ClassAttributeTest.send(:class_attributes)).to be_a(Hanami::Utils::ClassAttribute::Attributes)
+  end
+
+  it "preserves class attributes if module is included multiple times" do
+    expect(DoubleInclude.foo).to be(1)
   end
 end