+++ spec/unit/hanami/utils/escape_spec.rb
@@ -11,10 +11,10 @@ RSpec.describe Hanami::Utils::Escape do
     test_string = "<script>".encode(Encoding::UTF_8)
 
     string = begin
-               test_string.encode(encoding)
-             rescue StandardError
-               nil
-             end
+      test_string.encode(encoding)
+    rescue StandardError
+      nil
+    end
 
     result << encoding if !string.nil? && string != test_string
   end